Fix search creation and listing failing on Postgres
Same class of bug as the earlier sessions.rotation_counter fix:
SearchRow::top_k was declared Option<i64> in Rust, but the underlying
Postgres column is INTEGER (32-bit). sqlx's Postgres decoder rejects a
struct field wider than its column outright, while SQLite's untyped
storage accepted the mismatch silently on every test in this suite that
runs against AppState::for_tests() (SQLite). This broke both creating a
new search (POST /v1/search/face) and listing past ones
(GET /v1/search) unconditionally on the live Postgres deployment,
confirmed directly from its logs:

  error occurred while decoding column "top_k": mismatched types;
  Rust type `core::option::Option<i64>` (as SQL type `INT8`) is not
  compatible with SQL type `INT4`

Fixed by matching the field width to the actual column width. Added an
opt-in Postgres-backed regression test, mirroring the existing
postgres_session_smoke.rs/postgres_audit_chain_smoke.rs convention, so
this class of bug keeps getting caught before reaching production rather
than passing silently against the SQLite-backed suite that runs in CI.

+++ server/src/db/mod.rs
@@ -1136,7 +1136,13 @@ pub struct SearchRow {
     pub status: String,
     pub latitude: Option<f64>,
     pub longitude: Option<f64>,
-    pub top_k: Option<i64>,
+    // Matches the Postgres column's actual width (`INTEGER`, 32-bit) —
+    // same class of bug as `SessionRow::rotation_counter` (see
+    // `db/session.rs`): sqlx's Postgres decoder rejects a struct field
+    // wider than the column it reads from, which SQLite's untyped
+    // storage silently tolerates. This broke both creating a search and
+    // listing past ones on Postgres.
+    pub top_k: Option<i32>,
     pub started_at: Option<String>,
     pub completed_at: Option<String>,
     pub failure_code: Option<String>,
